Start refactoring for org wide Dangerfile
